Radios: Fix conditionally revealed content
