+++ components/radios/spec/Radios.stories.mdx
@@ -218,8 +218,6 @@ users if you separate them using a text divider. The text is usually the word
 
 ### Conditionally revealing content
 
-#### IMPLEMENT ME
-
 Using this component, you can add conditionally revealing content to stacked
 radios, so users only see content when it’s relevant to them.
